:new: Added get_quadrant and check_region
Added two functions:
get_quadrant: its take the position of a number and its returns the coordinates of the rows and columns of the quadrant that the number makes part.
check_region: its takes the actual states of the sudoku, a number and his actual position on it, and it returns true if this number already exist in the same quadrant, and false if not.